backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2022.4
sqlparse==0.4.3
